refactor : 상위 타입으로 추상 클래스 분리
