+++ src/main/java/domain/Dealer.java
@@ -1,7 +1,9 @@
 
 package domain;
 
-public class Dealer {
+import java.util.List;
+
+public class Dealer extends Gamer {
     private static final int HIT_CONDITION = 17;
     private final Hand hand;
 
@@ -9,13 +11,10 @@ public class Dealer {
         this.hand = new Hand();
     }
 
+    @Override
     public void hit(final Decks decks) {
         while (hand.sum() < HIT_CONDITION) {
             hand.add(decks.draw());
         }
-    }
-
-    public int calculateTotalScore() {
-        return hand.sum();
     }
 }
