+++ src/main/java/domain/Player.java
@@ -2,28 +2,21 @@ package domain;
 
 import java.util.List;
 
-public class Player {
+public class Player extends Gamer {
     private final Name name;
-    private final Hand hand;
+    //private final Hand hand;
 
     public Player(final Name name) {
         this.name = name;
-        this.hand = new Hand();
+        //this.hand = new Hand();
     }
 
-    public void hit(Decks decks) {
+    @Override
+    public void hit(final Decks decks) {
         hand.add(decks.draw());
-    }
-
-    public int calculateTotalScore() {
-        return hand.sum();
     }
 
     public Name getName() {
         return name;
-    }
-
-    public List<Card> getHand() {
-        return hand.getCards();
     }
 }
