+++ src/test/java/domain/DecksTest.java
@@ -2,6 +2,7 @@ package domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.strategy.ShuffledDecksGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -11,25 +12,13 @@ public class DecksTest {
     @Test
     void createDecksTest() {
         // given
-        int expectedSize = 6;
+        int expectedSize = 312;
 
         // when
-        Decks decks = new Decks();
+        ShuffledDecksGenerator decksGenerator = new ShuffledDecksGenerator();
+        Decks decks = Decks.createByStrategy(decksGenerator);
 
         // then
         assertThat(decks.getCards()).hasSize(expectedSize);
-    }
-
-    @DisplayName("모든 덱을 돌아가면서 카드를 뽑는다.")
-    @Test
-    void drawTest() {
-        // given
-        Decks decks = new Decks();
-
-        // when
-        decks.draw();
-
-        // then
-        assertThat(decks.getCards()).hasSize(51);
     }
 }
